requirements.txt
gensim==0.13.1
wget==3.2
mecab-python3==0.996.1
